Move native dumb component map into own file
