+++ shared/more/dumb-component-map.desktop.js
