+++ shared/test/render-dumb-sheet.js
@@ -5,7 +5,7 @@ import React, {Component} from 'react'
 import ReactDOM from 'react-dom'
 import {styleBox} from '../more/dumb-sheet.render.desktop'
 import {Box, Text} from '../common-adapters'
-import dumbComponentMap from '../more/dumb-component-map'
+import dumbComponentMap from '../more/dumb-component-map.desktop'
 
 const PADDING = 25
